requirements.txt
Django==4.1.3
djangorestframework==3.14.0
psycopg2
djangorestframework-simplejwt
openpyxl
python-dotenv